fix(profile): owner self-view RSC boundary — extract follow-stat helpers
ProfileView (Server Component) called followStatText() and used
FOLLOW_STAT_CLASS in the owner branch (isOwner === true). Both were exports of
FollowButton.tsx, a "use client" module — invoking a client-module function
from server render crosses the RSC boundary and throws, so every claimed
creator's own signed-in self-view rendered error.tsx ("Couldn't load that
creator."). Non-owner / visitor views were unaffected (they render
<FollowButton/> as a client island, where followStatText runs legally).

Option 1 (smallest correct): move followStatText + FOLLOW_STAT_CLASS verbatim
into a new directive-free module src/components/profile/follow-stat.ts, neutral
so both server and client can import it. ProfileView now imports the helpers
from there (legal server-side call); FollowButton imports them from the same
module (unchanged behavior). Single source of truth keeps the owner static
byline byte-identical to the client island's optimistic byline.

Render-layer only. No DB change, no migration, no follows/view touched.
tsc + next build clean.
